feat: enable hourly submission checking

+++ src/setupClient.ts
@@ -1,7 +1,9 @@
-import { Client, ThreadChannel } from 'discord.js'
+import { Client, ThreadChannel, User } from 'discord.js'
 
 import { getChannelMessages } from './getChannelMessages.js'
 import { error, info } from './log.js'
+
+const twoDaysAgo = Date.now() - 1000 * 60 * 60 * 24 * 2
 
 const setupChannel = (client: Client, channelId = ''): ThreadChannel | void => {
   const channel = client.channels.cache.get(channelId)
@@ -35,6 +37,8 @@ export const setupClient = async (client: Client) => {
     process.env.DISCORD_JUDGE_CHANNEL_ID
   )
 
+  const usersWithInvalidSubmissions = new Set<User['id']>()
+
   if (!discussionChannel || !submissionChannel || !judgeChannel) {
     // eslint-disable-next-line unicorn/no-process-exit
     process.exit(1)
@@ -42,8 +46,23 @@ export const setupClient = async (client: Client) => {
 
   // Delete any bot messages sent in a previous run
   for await (const message of getChannelMessages(discussionChannel)) {
-    if (message.author.bot) {
+    if (message.author.bot && message.createdTimestamp > twoDaysAgo) {
       info(`Deleting bot message ${message.id}`, import.meta, true)
+      message.delete()
+    } else if (message.author.bot) {
+      // don't ping user again if they've already been pinged in the last 2 days
+      // about their submission not being valid
+      const user = message.mentions.users.first()
+      if (!user) continue
+
+      usersWithInvalidSubmissions.add(user.id)
+    }
+  }
+
+  // Delete previous playlist attachments in the judge channel
+  for await (const message of getChannelMessages(judgeChannel)) {
+    if (message.author.bot && message.attachments.size > 0) {
+      info(`Deleting playlist ${message.id}`, import.meta, true)
       message.delete()
     }
   }
@@ -51,6 +70,7 @@ export const setupClient = async (client: Client) => {
   return {
     discussionChannel,
     submissionChannel,
-    judgeChannel
+    judgeChannel,
+    usersWithInvalidSubmissions
   }
 }
